engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.28.1
python-dotenv==1.0.1
scipy==1.17.1
numpy==2.4.4
apscheduler==3.10.4
